Exhaustive test suite, fixed bugs

+++ src/main/java/com/ecse437/project/math/CelsiusToFahrenheit.java
@@ -1,10 +1,14 @@
 package com.ecse437.project.math;
 
 public class CelsiusToFahrenheit {
-
+    /**
+     * Converts celsius to fahrenheit
+     * @param a value to be converted to fahrenheit
+     * @return converted fahrenheit value
+     */
     public int convert(int a) {
 
-        return (int) AddExact.values((long)(a*1.8), 32);
+        return (int) (long)AddExact.values((long)(a*1.8), 32);
     }
 }
